Split tests for different rule types
