+++ .betterer.results
@@ -2907,7 +2907,7 @@ exports[`better eslint`] = {
       [0, 0, 0, "Unexpected any. Specify a different type.", "2"],
       [0, 0, 0, "Unexpected any. Specify a different type.", "3"]
     ],
-    "public/app/features/alerting/unified/RuleEditor.test.tsx:5381": [
+    "public/app/features/alerting/unified/RuleEditorCloudRules.test.tsx:5381": [
       [0, 0, 0, "Unexpected any. Specify a different type.", "0"]
     ],
     "public/app/features/alerting/unified/RuleList.test.tsx:5381": [
